refactor: reorganize folder structure to align with FSD architecture
- Adjusted file locations for improved modularity and clarity
