+++ src/app/app.tsx
@@ -1,9 +1,9 @@
 import { Outlet } from "react-router-dom";
 import styles from "./app.module.scss";
 import { Suspense } from "react";
+import { ErrorBoundary } from "./providers/errorBoundary";
 import { useSearchModal } from "@contexts/searchModalContext";
 import DefaultLayout from "@layouts/defaultLayout";
-import { ErrorBoundary } from "@components/errorBoundary";
 
 export default function App() {
   const { setIsOpen } = useSearchModal().actions;
