+++ src/app/index.tsx
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom/client";
 import "@styles/base/index.scss";
 import { ApiProvider } from "@contexts/apiContext";
 import { SearchModalProvider } from "@contexts/searchModalContext";
-import Routes from "../shared/routes/routes";
+import Routes from "./routes/routes";
 
 async function enableMocking() {
   if (process.env.NODE_ENV !== "development") {
