+++ src/app/providers/errorBoundary/HTTP_ERROR_MESSAGE.tsx
