+++ src/shared/ui/customSelect/listModal/index.module.scss
