+++ src/shared/ui/searchPanel/index.tsx
