+++ src/shared/ui/searchPanel/searchBar/index.tsx
