+++ src/shared/ui/searchPanel/searchModal/index.tsx
