+++ src/widgets/footer/index.tsx
